refactor: 프로필 생성 책임을 ProfileGenerator로 이동
